feat: Add SPO resource to berkas.php
This commit adds the SPO resource to the berkas.php file in the routes/partials directory. The SPO resource is associated with the RsiaSpoController class and has the 'spo' parameter set to 'nomor'. This change allows for handling SPO-related functionality in the application.

Note: This suggested commit message is based on the provided code changes and recent repository commits.