:bug: fix: Update user context in RSIALogger to handle unknown user
The RSIALogger class in the app/Helpers/Logger/RSIALogger.php file has been updated to handle cases where the user is unknown. Previously, if the user was not authenticated, an error would occur when trying to access the user's ID. Now, if the user is not authenticated, the 'user' context in the logger will be set to 'unknown'. This change ensures that the logger functions correctly even when the user is not logged in.