:recycle: refactor: Add PenerimaUndanganWithKehadiranCollection resource for transforming the resource collection
This commit adds a new file, PenerimaUndanganWithKehadiranCollection.php, which is a resource collection class used for transforming the collection of PenerimaUndanganWithKehadiran resources. The toArray() method in this class maps each item in the collection and merges additional data with the existing resource attributes. This change improves the structure and presentation of the transformed data.