:hammer: update : Add exposedScopes method to BookingRegistrasiController
The BookingRegistrasiController.php file has been updated to include a new method called exposedScopes. This method returns an array of available query scopes, with the addition of the 'statusBelum' scope. This change enhances the functionality of the controller by providing a convenient way to access and use the 'statusBelum' scope.