feat: Add 'tidakBatal' scope to BookingRegistrasiController
The `BookingRegistrasiController` has been updated to include a new scope method called `tidakBatal`. This scope allows filtering the `BookingRegistrasi` records that do not have a status of 'Batal'. This change enhances the functionality of the controller by providing a convenient way to query and retrieve specific records based on their status.

Co-authored-by: Faisal Halim <ffaisalhalim@gmail.com>