:sparkles: feat: Generate no_reg based on lastRegistrasi in BookingRegistrasiController
The `BookingRegistrasiController.php` file has been updated to generate the `no_reg` value based on the `lastRegistrasi` retrieved from the `RegPeriksa` model. If `lastRegistrasi` exists, the `no_rawat` value is extracted and incremented to generate the `no_reg` value. Otherwise, `no_reg` is set to 0. This change improves the accuracy of the `no_reg` value for each booking registration.