refactor: Include 'no_rkm_medis' filter in RegPeriksaController
The RegPeriksaController.php file has been updated to include the 'no_rkm_medis' filter in the 'filterableBy' method. This change allows filtering 'RegPeriksa' records based on the 'no_rkm_medis' field. It improves the functionality of the controller by providing the ability to search for records using the patient's medical record number.