feat: Add RsiaTemporaryPresensiController with index method
A new controller, RsiaTemporaryPresensiController, has been added with an index method. This method retrieves the temporary presence data for a specific employee based on their NIK. If the employee is not found, an error response is returned. The controller is responsible for fetching the data from the TemporaryPresensi model and returning it as a RealDataResource.