:sparkles: feat : PresensiKaryawanController to include sorting and filtering attributes
The PresensiKaryawanController.php file has been updated to include the sortableBy() and filterableBy() methods. These methods define the attributes that can be used for sorting and filtering the presensi data. This change enhances the functionality of the controller by allowing users to sort and filter the presensi records based on various attributes.