🛡️ feat (middleware) : Add CustomUserMiddleware to handle dynamic user authentication
This commit adds a new middleware class, CustomUserMiddleware, to handle dynamic user authentication. The middleware checks for an authenticated user in each guard defined in the auth configuration. If a user is found, it sets the user in the request and updates the user resolver. This allows for flexible user authentication based on the active guard.