:hammer: Update: ProsedurPasienController to include filterable and sortable attributes
This commit updates the ProsedurPasienController to include the filterableBy and sortableBy attributes. These attributes define the fields that can be used for filtering and sorting the ProsedurPasien resources. The filterableBy attribute includes the fields 'no_rawat', 'kode', 'status', and 'prioritas', while the sortableBy attribute includes the fields 'no_rawat', 'kd_penyakit', 'status', and 'prioritas'. This enhancement improves the flexibility and usability of the ProsedurPasienController.