:hammer: Update: RsiaSpoController to use updateOrCreate for detail data
This commit refactors the RsiaSpoController by updating the `afterUpdate` method to use the `updateOrCreate` method instead of `update` for the `detail` relationship. This change ensures that the detail data is either updated if it exists or created if it doesn't exist, based on the `nomor` field. This improves the reliability and consistency of the data handling in the controller.