:sparkles: feat: Add JasaPegawaiController for handling jasa pegawai data
This commit adds a new controller, JasaPegawaiController, to handle jasa pegawai data. It includes two methods, `jm` and `jaspel`, for retrieving jasa medis and jasa pelayanan data respectively. These methods fetch the data from the corresponding models and return it as a paginated collection. This addition enhances the functionality of the application by providing endpoints to retrieve jasa pegawai data.