:recycle: Refactor (BridgingSep) : Add mass assignable attributes and relationships
This commit refactors the BridgingSep model by adding mass assignable attributes and relationships. The `scopeGetTanggalPulang` method is added to retrieve the tanggal_pulang attribute based on the given noRawat. The `status_klaim` relationship is added to retrieve the status_klaim that owns the BridgingSep. The `dokter` relationship is added to retrieve the dokter that owns the BridgingSep. The `tanggal_pulang` relationship is added to retrieve the tanggal_pulang that owns the BridgingSep.