:sparkles: feat: Add 'sync' method to KlaimController
This commit adds a new method called 'sync' to the KlaimController. The 'sync' method is responsible for retrieving klaim data from the eklaim service and saving it to our database. It uses the BodyBuilder class to set metadata and data for the request, sends the request to the eklaim service, and handles the response. If the response is successful, the klaim data is inserted into the inacbg_grouping_stage12 table in the database. If there is an error, it is logged and an error response is returned.

The 'sync' method is added to the routes in the bridging-sep.php file.

This commit addresses the need to synchronize klaim data between our system and the eklaim service.